StippleUI.Separators.separator — Function

  separator(args...; kwargs...)

The separator component is used to separate sections of text, other components, etc… It creates a colored line across the width of the parent. It can be horizontal or vertical.